refactor: generalized cleanup and refactoring

+++ src/ExperimentFramework.Science/EffectSize/OddsRatio.cs
@@ -68,7 +68,7 @@ public sealed class OddsRatio : IBinaryEffectSizeCalculator
         else
         {
             // Cast both operands before multiplication to avoid potential integer overflow
-            or = ((double)a * (double)d) / ((double)b * (double)c);
+            or = (a * (double)d) / (b * (double)c);
             logOrSe = Math.Sqrt(1.0 / a + 1.0 / b + 1.0 / c + 1.0 / d);
         }
